Fix: Preserve negative sign in number input on nested x-model Alpine's x-model was not being updated when the element value was changing, causing a bug when trying to save negative numbers. I ran into the issue when I was binding my input like so: <input x-model="values.value" x-number> where values is just an object: values = { value: 123 }
When I have a @click handler to do an ajax submit, the value or values.value would not change if only the sign changed.

This fix triggers alpinejs' event handling to update the model value.

Note: I ran `npm run build` after making the change, which I assume is correct to get the dist files.